test: create tests to entity models

+++ backend/src/test/java/com/mini/agro/backend/model/entity/ProductionRecordTest.java
@@ -1,10 +1,10 @@
 package com.mini.agro.backend.model.entity;
 
-import com.mini.agro.backend.model.entity.Commodity;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 class ProductionRecordTest {
@@ -25,28 +25,28 @@ class ProductionRecordTest {
 
     @Test
     void setPlantsByHectare() {
-        Long plantsByHectareTest = 1L;
+        BigDecimal plantsByHectareTest = BigDecimal.valueOf(1);
         productionRecord.setPlantsByHectare(plantsByHectareTest);
         Assertions.assertEquals(Optional.of(plantsByHectareTest), Optional.ofNullable(productionRecord.getPlantsByHectare()));
     }
 
     @Test
     void setPodsByPlant() {
-        Long podsByPlantTest = 1L;
+        BigDecimal podsByPlantTest = BigDecimal.valueOf(1);
         productionRecord.setPodsByPlant(podsByPlantTest);
         Assertions.assertEquals(Optional.of(podsByPlantTest), Optional.ofNullable(productionRecord.getPodsByPlant()));
     }
 
     @Test
     void setGrainsByPod() {
-        Long grainsByPodTest = 1L;
+        BigDecimal grainsByPodTest = BigDecimal.valueOf(1);
         productionRecord.setGrainsByPod(grainsByPodTest);
         Assertions.assertEquals(Optional.of(grainsByPodTest), Optional.ofNullable(productionRecord.getGrainsByPod()));
     }
 
     @Test
     void setWeightByOneKGrains() {
-        Long weightByOneKGrainsTest = 1L;
+        BigDecimal weightByOneKGrainsTest = BigDecimal.valueOf(1);
         productionRecord.setWeightByOneKGrains(weightByOneKGrainsTest);
         Assertions.assertEquals(Optional.of(weightByOneKGrainsTest), Optional.ofNullable(productionRecord.getWeightByOneKGrains()));
     }
@@ -79,10 +79,10 @@ class ProductionRecordTest {
     private ProductionRecord createProductionRecord() {
         ProductionRecord productionRecord = new ProductionRecord();
         productionRecord.setCommodity(new Commodity());
-        productionRecord.setPlantsByHectare(1L);
-        productionRecord.setPodsByPlant(1L);
-        productionRecord.setGrainsByPod(1L);
-        productionRecord.setWeightByOneKGrains(1L);
+        productionRecord.setPlantsByHectare(BigDecimal.valueOf(1));
+        productionRecord.setPodsByPlant(BigDecimal.valueOf(1));
+        productionRecord.setGrainsByPod(BigDecimal.valueOf(1));
+        productionRecord.setWeightByOneKGrains(BigDecimal.valueOf(1));
         return productionRecord;
     }
